+++ static/render.js
@@ -1,0 +1,9 @@
+(function () {
+  function setEmptyHtml(element, html) {
+    if (element) {
+      element.innerHTML = html;
+    }
+  }
+
+  window.AeroTravelRender = Object.freeze({ setEmptyHtml });
+})();